feat: 핸들러 어댑터 일급 콜렉션 추가
